refactor token util into dedicated module
